:recycle: Refactor the code to extra PEvalBitwiseTerm class
