+++ grisette.cabal
@@ -102,11 +102,12 @@ library
       Grisette.IR.SymPrim.Data.IntBitwidth
       Grisette.IR.SymPrim.Data.Prim.Helpers
       Grisette.IR.SymPrim.Data.Prim.Internal.Caches
+      Grisette.IR.SymPrim.Data.Prim.Internal.Instances.PEvalBitwiseTerm
+      Grisette.IR.SymPrim.Data.Prim.Internal.Instances.SupportedPrim
       Grisette.IR.SymPrim.Data.Prim.Internal.Term
       Grisette.IR.SymPrim.Data.Prim.Model
       Grisette.IR.SymPrim.Data.Prim.ModelValue
       Grisette.IR.SymPrim.Data.Prim.PartialEval.Bits
-      Grisette.IR.SymPrim.Data.Prim.PartialEval.Bool
       Grisette.IR.SymPrim.Data.Prim.PartialEval.BV
       Grisette.IR.SymPrim.Data.Prim.PartialEval.Integral
       Grisette.IR.SymPrim.Data.Prim.PartialEval.Num
