+++ src/Grisette/Core/Data/Class/ITEOp.hs
@@ -18,10 +18,9 @@ module Grisette.Core.Data.Class.ITEOp
 where
 
 import GHC.TypeNats (KnownNat, type (<=))
-import Grisette.IR.SymPrim.Data.Prim.PartialEval.Bool (pevalITETerm)
 import Grisette.IR.SymPrim.Data.Prim.Term
   ( LinkedRep,
-    SupportedPrim,
+    SupportedPrim (pevalITETerm),
   )
 import Grisette.IR.SymPrim.Data.SymPrim
   ( SymBool (SymBool),
