+++ src/Grisette/Internal/IR/SymPrim.hs
@@ -49,7 +49,6 @@ module Grisette.Internal.IR.SymPrim
     pevalNotEqvTerm,
     pevalOrTerm,
     pevalAndTerm,
-    pevalITETerm,
     pevalImplyTerm,
     pevalXorTerm,
     unaryUnfoldOnce,
@@ -95,22 +94,6 @@ import Grisette.IR.SymPrim.Data.Prim.Helpers
     pattern UnaryTermPatt,
   )
 import Grisette.IR.SymPrim.Data.Prim.Model (evaluateTerm)
-import Grisette.IR.SymPrim.Data.Prim.PartialEval.Bool
-  ( falseTerm,
-    pevalAndTerm,
-    pevalEqvTerm,
-    pevalITETerm,
-    pevalImplyTerm,
-    pevalNotEqvTerm,
-    pevalNotTerm,
-    pevalOrTerm,
-    pevalXorTerm,
-    trueTerm,
-    pattern BoolConTerm,
-    pattern BoolTerm,
-    pattern FalseTerm,
-    pattern TrueTerm,
-  )
 import Grisette.IR.SymPrim.Data.Prim.PartialEval.Integral
   ( pevalDivIntegralTerm,
     pevalModIntegralTerm,
@@ -164,16 +147,29 @@ import Grisette.IR.SymPrim.Data.Prim.Term
     constructBinary,
     constructTernary,
     constructUnary,
+    falseTerm,
     identity,
     identityWithTypeRep,
     introSupportedPrimConstraint,
     isymTerm,
+    pevalAndTerm,
+    pevalEqvTerm,
+    pevalImplyTerm,
+    pevalNotEqvTerm,
+    pevalNotTerm,
+    pevalOrTerm,
+    pevalXorTerm,
     pformat,
     showUntyped,
     someTypedSymbol,
     ssymTerm,
     symTerm,
+    trueTerm,
     withSymbolSupported,
+    pattern BoolConTerm,
+    pattern BoolTerm,
+    pattern FalseTerm,
+    pattern TrueTerm,
   )
 import Grisette.IR.SymPrim.Data.Prim.TermUtils
   ( castTerm,
