+++ test/Grisette/Core/Data/Class/SEqTests.hs
@@ -27,7 +27,7 @@ import Grisette.Core.Data.Class.LogicalOp (LogicalOp (symNot, (.&&)))
 import Grisette.Core.Data.Class.SEq (SEq ((./=), (.==)))
 import Grisette.Core.Data.Class.Solvable (Solvable (con))
 import Grisette.Core.Data.Class.TestValues (conBool, ssymBool)
-import Grisette.IR.SymPrim.Data.Prim.PartialEval.Bool (pevalEqvTerm)
+import Grisette.IR.SymPrim.Data.Prim.Term (pevalEqvTerm)
 import Grisette.IR.SymPrim.Data.SymPrim (SymBool (SymBool))
 import Test.Framework (Test, testGroup)
 import Test.Framework.Providers.HUnit (testCase)
@@ -101,12 +101,12 @@ seqTests =
                     [ssymBool "a"]
                       .== [ssymBool "b"]
                       @=? ssymBool "a"
-                        .== ssymBool "b",
+                      .== ssymBool "b",
                   testCase "Same length 2" $
                     [ssymBool "a", ssymBool "b"]
                       .== [ssymBool "c", ssymBool "d"]
                       @=? (ssymBool "a" .== ssymBool "c")
-                        .&& (ssymBool "b" .== ssymBool "d"),
+                      .&& (ssymBool "b" .== ssymBool "d"),
                   testCase "length 1 vs length 0" $
                     [ssymBool "a"] .== [] @=? conBool False,
                   testCase "length 1 vs length 2" $
@@ -131,7 +131,7 @@ seqTests =
                     Just (ssymBool "a")
                       .== Just (ssymBool "b")
                       @=? ssymBool "a"
-                        .== ssymBool "b"
+                      .== ssymBool "b"
                 ]
             ],
           testGroup
@@ -144,7 +144,7 @@ seqTests =
                     (Left (ssymBool "a") :: Either SymBool SymBool)
                       .== Left (ssymBool "b")
                       @=? ssymBool "a"
-                        .== ssymBool "b",
+                      .== ssymBool "b",
                   testCase "Right vs Left" $
                     (Right (ssymBool "a") :: Either SymBool SymBool)
                       .== Left (ssymBool "b")
@@ -157,7 +157,7 @@ seqTests =
                     (Right (ssymBool "a") :: Either SymBool SymBool)
                       .== Right (ssymBool "b")
                       @=? ssymBool "a"
-                        .== ssymBool "b"
+                      .== ssymBool "b"
                 ]
             ],
           testGroup
@@ -208,7 +208,7 @@ seqTests =
                               MaybeT Maybe SymBool
                           )
                       @=? ssymBool "a"
-                        .== ssymBool "b"
+                      .== ssymBool "b"
                 ]
             ],
           testGroup
@@ -246,7 +246,7 @@ seqTests =
                               ExceptT SymBool Maybe SymBool
                           )
                       @=? ssymBool "a"
-                        .== ssymBool "b",
+                      .== ssymBool "b",
                   testCase
                     "ExceptT (Just (Right v)) vs ExceptT (Just (Left v))"
                     $ ExceptT (Just (Right (ssymBool "a")))
@@ -268,7 +268,7 @@ seqTests =
                               ExceptT SymBool Maybe SymBool
                           )
                       @=? ssymBool "a"
-                        .== ssymBool "b"
+                      .== ssymBool "b"
                 ]
             ],
           testProperty "()" (ioProperty . concreteSEqOkProp @()),
@@ -280,9 +280,9 @@ seqTests =
                 (ssymBool "a", ssymBool "c")
                   .== (ssymBool "b", ssymBool "d")
                   @=? ssymBool "a"
-                    .== ssymBool "b"
-                    .&& ssymBool "c"
-                      .== ssymBool "d"
+                  .== ssymBool "b"
+                  .&& ssymBool "c"
+                  .== ssymBool "d"
             ],
           testGroup
             "(,,)"
@@ -292,9 +292,9 @@ seqTests =
                 (ssymBool "a", ssymBool "c", ssymBool "e")
                   .== (ssymBool "b", ssymBool "d", ssymBool "f")
                   @=? (ssymBool "a" .== ssymBool "b")
-                    .&& ( (ssymBool "c" .== ssymBool "d")
-                            .&& (ssymBool "e" .== ssymBool "f")
-                        )
+                  .&& ( (ssymBool "c" .== ssymBool "d")
+                          .&& (ssymBool "e" .== ssymBool "f")
+                      )
             ],
           testGroup
             "(,,,)"
@@ -308,9 +308,9 @@ seqTests =
                   @=? ( (ssymBool "a" .== ssymBool "b")
                           .&& (ssymBool "c" .== ssymBool "d")
                       )
-                    .&& ( (ssymBool "e" .== ssymBool "f")
-                            .&& (ssymBool "g" .== ssymBool "h")
-                        )
+                  .&& ( (ssymBool "e" .== ssymBool "f")
+                          .&& (ssymBool "g" .== ssymBool "h")
+                      )
             ],
           testGroup
             "(,,,,)"
@@ -335,11 +335,11 @@ seqTests =
                   @=? ( (ssymBool "a" .== ssymBool "b")
                           .&& (ssymBool "c" .== ssymBool "d")
                       )
-                    .&& ( (ssymBool "e" .== ssymBool "f")
-                            .&& ( (ssymBool "g" .== ssymBool "h")
-                                    .&& (ssymBool "i" .== ssymBool "j")
-                                )
-                        )
+                  .&& ( (ssymBool "e" .== ssymBool "f")
+                          .&& ( (ssymBool "g" .== ssymBool "h")
+                                  .&& (ssymBool "i" .== ssymBool "j")
+                              )
+                      )
             ],
           testGroup
             "(,,,,,)"
@@ -375,11 +375,11 @@ seqTests =
                                   .&& (ssymBool "e" .== ssymBool "f")
                               )
                       )
-                    .&& ( (ssymBool "g" .== ssymBool "h")
-                            .&& ( (ssymBool "i" .== ssymBool "j")
-                                    .&& (ssymBool "k" .== ssymBool "l")
-                                )
-                        )
+                  .&& ( (ssymBool "g" .== ssymBool "h")
+                          .&& ( (ssymBool "i" .== ssymBool "j")
+                                  .&& (ssymBool "k" .== ssymBool "l")
+                              )
+                      )
             ],
           testGroup
             "(,,,,,,)"
@@ -419,13 +419,13 @@ seqTests =
                                     .&& (ssymBool "e" .== ssymBool "f")
                                 )
                         )
-                      .&& ( ( (ssymBool "g" .== ssymBool "h")
-                                .&& (ssymBool "i" .== ssymBool "j")
-                            )
-                              .&& ( (ssymBool "k" .== ssymBool "l")
-                                      .&& (ssymBool "m" .== ssymBool "n")
-                                  )
+                    .&& ( ( (ssymBool "g" .== ssymBool "h")
+                              .&& (ssymBool "i" .== ssymBool "j")
                           )
+                            .&& ( (ssymBool "k" .== ssymBool "l")
+                                    .&& (ssymBool "m" .== ssymBool "n")
+                                )
+                        )
             ],
           testGroup
             "(,,,,,,,)"
@@ -469,13 +469,13 @@ seqTests =
                                   .&& (ssymBool "g" .== ssymBool "h")
                               )
                       )
-                    .&& ( ( (ssymBool "i" .== ssymBool "j")
-                              .&& (ssymBool "k" .== ssymBool "l")
-                          )
-                            .&& ( (ssymBool "m" .== ssymBool "n")
-                                    .&& (ssymBool "o" .== ssymBool "p")
-                                )
+                  .&& ( ( (ssymBool "i" .== ssymBool "j")
+                            .&& (ssymBool "k" .== ssymBool "l")
                         )
+                          .&& ( (ssymBool "m" .== ssymBool "n")
+                                  .&& (ssymBool "o" .== ssymBool "p")
+                              )
+                      )
             ],
           testCase "ByteString" $ do
             let bytestrings :: [B.ByteString] = ["", "a", "ab"]
@@ -500,7 +500,7 @@ seqTests =
                     (InL $ Just $ ssymBool "a" :: Sum Maybe Maybe SymBool)
                       .== InL (Just $ ssymBool "b")
                       @=? ssymBool "a"
-                        .== ssymBool "b",
+                      .== ssymBool "b",
                   testCase "InL (Just v) vs InR (Just v)" $
                     (InL $ Just $ ssymBool "a" :: Sum Maybe Maybe SymBool)
                       .== InR (Just $ ssymBool "b")
@@ -509,7 +509,7 @@ seqTests =
                     (InR $ Just $ ssymBool "a" :: Sum Maybe Maybe SymBool)
                       .== InR (Just $ ssymBool "b")
                       @=? ssymBool "a"
-                        .== ssymBool "b",
+                      .== ssymBool "b",
                   testCase "InR (Just v) vs InL (Just v)" $
                     (InR $ Just $ ssymBool "a" :: Sum Maybe Maybe SymBool)
                       .== InL (Just $ ssymBool "b")
@@ -539,7 +539,7 @@ seqTests =
                         )
                           .== WriterLazy.WriterT (Left $ ssymBool "b")
                           @=? ssymBool "a"
-                            .== ssymBool "b",
+                          .== ssymBool "b",
                       testCase "WriterT (Left v) vs WriterT (Right v)" $
                         ( WriterLazy.WriterT (Left $ ssymBool "a") ::
                             WriterLazy.WriterT SymBool (Either SymBool) SymBool
@@ -562,7 +562,7 @@ seqTests =
                           .== WriterLazy.WriterT
                             (Right (ssymBool "c", ssymBool "d"))
                           @=? (ssymBool "a" .== ssymBool "c")
-                            .&& (ssymBool "b" .== ssymBool "d")
+                          .&& (ssymBool "b" .== ssymBool "d")
                     ]
                 ],
               testGroup
@@ -589,7 +589,7 @@ seqTests =
                         )
                           .== WriterStrict.WriterT (Left $ ssymBool "b")
                           @=? ssymBool "a"
-                            .== ssymBool "b",
+                          .== ssymBool "b",
                       testCase "WriterT (Left v) vs WriterT (Right v)" $
                         ( WriterStrict.WriterT (Left $ ssymBool "a") ::
                             WriterStrict.WriterT
@@ -621,9 +621,9 @@ seqTests =
                           .== WriterStrict.WriterT
                             (Right (ssymBool "c", ssymBool "d"))
                           @=? ssymBool "a"
-                            .== ssymBool "c"
-                            .&& ssymBool "b"
-                              .== ssymBool "d"
+                          .== ssymBool "c"
+                          .&& ssymBool "b"
+                          .== ssymBool "d"
                     ]
                 ]
             ],
@@ -638,7 +638,7 @@ seqTests =
                 (Identity $ ssymBool "a" :: Identity SymBool)
                   .== Identity (ssymBool "b")
                   @=? ssymBool "a"
-                    .== ssymBool "b"
+                  .== ssymBool "b"
             ],
           testGroup
             "IdentityT"
@@ -655,7 +655,7 @@ seqTests =
                     )
                       .== IdentityT (Left $ ssymBool "b")
                       @=? ssymBool "a"
-                        .== ssymBool "b",
+                      .== ssymBool "b",
                   testCase "IdentityT (Left v) vs IdentityT (Right v)" $
                     ( IdentityT $ Left $ ssymBool "a" ::
                         IdentityT (Either SymBool) SymBool
@@ -674,7 +674,7 @@ seqTests =
                     )
                       .== IdentityT (Right $ ssymBool "b")
                       @=? ssymBool "a"
-                        .== ssymBool "b"
+                      .== ssymBool "b"
                 ]
             ]
         ],
@@ -694,7 +694,7 @@ seqTests =
                 A2 (ssymBool "a")
                   .== A2 (ssymBool "b")
                   @=? ssymBool "a"
-                    .== ssymBool "b",
+                  .== ssymBool "b",
               testCase "A2 vs A3" $
                 A2 (ssymBool "a")
                   .== A3 (ssymBool "b") (ssymBool "c")
@@ -709,7 +709,7 @@ seqTests =
                 A3 (ssymBool "a") (ssymBool "b")
                   .== A3 (ssymBool "c") (ssymBool "d")
                   @=? (ssymBool "a" .== ssymBool "c")
-                    .&& (ssymBool "b" .== ssymBool "d")
+                  .&& (ssymBool "b" .== ssymBool "d")
             ]
         ]
     ]
